+++ src/Proxy/DriveProxy.php
@@ -121,6 +121,42 @@ class DriveProxy extends BaseItemProxy
     }
 
     /**
+     * Gets a drive item by path from this instance.
+     *
+     * @param string $path
+     *        The path.
+     *
+     * @return DriveItemProxy
+     *         The drive item.
+     *
+     * @since 2.2.0
+     *
+     * @link https://docs.microsoft.com/en-us/onedrive/developer/rest-api/api/driveitem_get?view=odsp-graph-online
+     *       Get a DriveItem resource
+     */
+    public function getDriveItemByPath($path)
+    {
+        $driveLocator = "/drives/{$this->id}";
+        $itemLocator  = "/root:$path";
+        $endpoint     = "$driveLocator$itemLocator";
+
+        $response = $this
+            ->graph
+            ->createRequest('GET', $endpoint)
+            ->execute();
+
+        $status = $response->getStatus();
+
+        if ($status != 200) {
+            throw new \Exception("Unexpected status code produced by 'GET $endpoint': $status");
+        }
+
+        $driveItem = $response->getResponseAsObject(DriveItem::class);
+
+        return new DriveItemProxy($this->graph, $driveItem);
+    }
+
+    /**
      * Gets the root of this instance.
      *
      * @return DriveItemProxy